refactor: fix lint due to golang-ci bump
