+++ internal/app/app_types.go
@@ -10,7 +10,7 @@ import (
 	"github.com/y3owk1n/neru/internal/config"
 	"github.com/y3owk1n/neru/internal/core/domain"
 	"github.com/y3owk1n/neru/internal/core/domain/state"
-	"github.com/y3owk1n/neru/internal/core/infra/metrics"
+	"github.com/y3owk1n/neru/internal/core/infra/appmetrics"
 	"github.com/y3owk1n/neru/internal/core/ports"
 	"github.com/y3owk1n/neru/internal/ui"
 	"go.uber.org/zap"
@@ -49,7 +49,7 @@ type App struct {
 	eventTap       ports.EventTapPort
 	ipcServer      ports.IPCPort
 	appWatcher     Watcher
-	metrics        metrics.Collector
+	metrics        appmetrics.Collector
 
 	modes *modes.Handler
