+++ internal/core/infra/appmetrics/doc.go
@@ -1,5 +1,5 @@
-// Package metrics provides metrics collection and reporting.
+// Package appmetrics provides metrics collection and reporting.
 //
 // This package implements a simple in-memory metrics collector for counters,
 // gauges, and histograms, with snapshot capabilities for external reporting.
-package metrics
+package appmetrics
