+++ internal/core/infra/appmetrics/metrics.go
@@ -1,4 +1,4 @@
-package metrics
+package appmetrics
 
 import (
 	"sync"
