+++ internal/core/infra/apptrace/doc.go
@@ -1,5 +1,5 @@
-// Package trace provides tracing utilities.
+// Package apptrace provides tracing utilities.
 //
 // This package implements unique trace ID generation and context propagation
 // to enable distributed tracing across the application.
-package trace
+package apptrace
